implement dup2, get lseek to report proper seeking behavior
lseek would return a huge non-negative integer when 0 was meant to be returned
when setting the offset to 0 with SEEK_SET. This is because it was returning an off_t,
which doesn't fit in the MIPS 32-bit registers. To compromise, we don't allow files
larger than 16GB, as the offset is now a int32_t, which fits fine. It should really be a
uint32_t, but lots of old code in the system casts to off_t, which is signed 64 bit, so it's safe to cast 32-bit signed to 64-bit signed.

Add testbin/luketest to test open, close, lseek and dup and dup2 in the future. This is just a playground file for testing various things.